SCRUM-130: Add content to user page and integrate report and admin tools on the page.
