+++ wellness-app/src/assets/Auth/ForgotPassword.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { auth } from './Firebase'; // Adjust the path if needed
+import { auth } from '../../Firebase'; // Adjust the path if needed
 import { sendPasswordResetEmail } from 'firebase/auth';
 import styles from '../styles/ForgotPassword.module.css';
