+++ wellness-app/src/assets/Auth/Login.jsx
@@ -3,9 +3,9 @@ import { Link, useNavigate } from "react-router-dom"; // Keep only one import st
 import "../styles/login.css";
 import "../styles/guide.css";
 import { FaUser, FaLock } from "react-icons/fa";
-import {auth} from "./Firebase";
+import {auth} from "../../Firebase";
 import {signInWithEmailAndPassword} from 'firebase/auth'
-import UserContext from "./UserContext";
+import UserContext from "../../UserContext";
 
 
 function Login() {
