+++ wellness-app/src/assets/Footer.jsx
@@ -14,8 +14,8 @@ function Footer(){
         <h4>Help</h4>
         <a href="/support">Support</a>
         <a href="#">Payment Options</a>
-        <a href="#">Privacy Policies</a>
-        <a href="/tos">Terms of service</a>
+        <a href="/privacy">Privacy Policy</a>
+        <a href="/tos">Terms Of Service</a>
       </div>
       <div className="footer-section">
         <h4>Other</h4>
