+++ wellness-app/src/assets/PrivacyPolicy.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import styles from '../styles/Legal.module.css'; // Adjust the path as necessary
+
+function PrivacyPolicy() {
+    return (
+        <div className={styles.legalContainer}>
+            <h1 className={styles.legalTitle}>Privacy Policy</h1>
+            <p className={styles.effectiveDate}><strong>Effective Date:</strong> [Insert Date]</p>
+
+            <p className={styles.introduction}>
+
+            </p>
+
+            <p className={styles.agreement}>
+
+            </p>
+
+            <h2 className={styles.sectionTitle}> </h2>
+            <p className={styles.sectionContent}>
+                // TODO: once Privacy Policy is finalized
+            </p>
+        </div>
+    );
+}
+
+export default PrivacyPolicy;
