+++ wellness-app/src/assets/Profile/ProfileText.jsx
