+++ wellness-app/src/assets/Ticket/Ticket.jsx
