+++ wellness-app/src/assets/Ticket/TicketItem.js
