+++ wellness-app/src/assets/Ticket/ViewTicket.js
