+++ wellness-app/src/styles/Legal.module.css
@@ -1,5 +1,5 @@
-/* TOS.module.css */
-.tosContainer {
+/* legal.module.css */
+.legalContainer {
     padding: 20px;
     width: 80%;
     background-color: #f4f4f4; /* Light grey background for contrast */
@@ -11,7 +11,7 @@
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1); /* Subtle shadow */
 }
 
-.tosTitle {
+.legalTitle {
     color: #0056b3; /* Dark blue color for the main heading */
     margin-bottom: 10px; /* Space below the title */
 }
