--- wellness-app/src/styles/UserPage.module.css
@@ -1,126 +1,0 @@
-.userPage {
-    font-family: Arial, sans-serif;
-    color: #333;
-}
-
-/* Banner */
-.banner {
-    height: 250px;
-    background-color: #f0f0f0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    margin-bottom: 20px; /* Add margin to separate it from navigation */
-}
-
-.profileImageWrapper {
-    position: absolute;
-    bottom: -30px;
-    width: 80px;
-    height: 80px;
-    border-radius: 50%;
-    overflow: hidden;
-    border: 2px solid #ddd;
-    background-color: white; /* Optional: to make the profile image stand out */
-}
-
-.profileImage {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-/* Navigation Links */
-.navLinks {
-    margin-top: 80px; /* Adjust to give space under the profile image */
-    display: flex;
-    justify-content: center;
-}
-
-.navButton {
-    background-color: transparent;
-    border: none;
-    padding: 10px 20px;
-    cursor: pointer;
-    font-size: 16px;
-    color: #007bff;
-    border-bottom: 2px solid transparent;
-    transition: border-color 0.3s;
-}
-
-.navButton:hover,
-.active {
-    border-color: #007bff;
-}
-
-/* Content Area */
-.contentArea {
-    padding: 20px;
-}
-
-/* Posts Filter Dropdown */
-.dropdown {
-    margin-bottom: 20px;
-}
-
-.dropdownSelect {
-    padding: 5px;
-    font-size: 14px;
-}
-
-/* Posts Placeholder */
-.postsContainer {
-    padding: 15px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    margin-bottom: 20px;
-}
-
-/* Shop section */
-.shopSection {
-    padding: 20px;
-}
-
-.productGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
-    gap: 20px;
-}
-
-.productCard {
-    border: 1px solid #ccc;
-    padding: 10px;
-    text-align: center;
-    cursor: pointer;
-}
-
-.productImage {
-    width: 100%;
-    height: auto;
-}
-
-.productName {
-    font-weight: bold;
-    margin-top: 10px;
-}
-
-.productPrice {
-    color: green;
-}
-
-/* About, Contact, Report, and ModView Sections */
-.aboutSection,
-.contactSection,
-.reportSection,
-.modSection {
-    padding: 15px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-}
-
-/* Links */
-.link {
-    color: #0056b3; /* Blue color for links */
-    text-decoration: none; /* Remove underline */
-}
